[REFACTOR] refactored tests in relation to element_objects restructuring
